Problem 7.Fix emails
You are given a sequence of strings, each on a new line, unitll you receive the "stop" command. The first string is a name of a person. On the second line you will receive his email. Your task is to collect their names and emails, and remove the entries whose email’s domain ends with "us" or "uk" (case insensitive). 
Print all the collected people and their emails in the following format: “{name} –> {email}”