Problem 1.Unique Usernames
On the first line you will be given an integer N. On the next N lines you will receive one username per line. Write a simple program that reads from the console a sequence of N usernames and keeps a collection only of the unique ones. Print the collection on the console in order of insertion: